feat(types): ✨ expose i64 as a surface type — full vertical (#136)
## Summary

Surfaces `i64` as a first-class type across the entire compiler stack.
Also adds integer literal contextual typing — `let x: i64 = 42` now
works without suffix syntax, with context propagated through let
initializers, assignments, call arguments, return statements, and binary
op peer types.

This completes Task 14 Tier B's primary deliverable: i64 surface
exposure as a Phase 6 prerequisite.

## Highlights

- **Contextual literal typing**: integer/float literals adopt the
expected type from declaration context, enabling `let x: i64 = 42`,
`return 0` in i64 functions, `add_i64(x, 1)`, and `x + 1` where x is i64
- **Runtime hooks**: `__dao_eq_i64` and `__dao_conv_i64_to_string`
(contract, C impl, LLVM declarations)
- **Widen `__dao_str_length`** from i32 to i64 across all four layers
(contract, runtime, LLVM decl, stdlib)
- **Stdlib**: i64 extends for Equatable, Printable, Numeric concepts
- **Example**: `factorial_i64(20)` = 2432902008176640000 — demonstrates
i64 arithmetic beyond i32 range
- **Checked overflow**: works automatically for i64 via existing
width-generic overflow intrinsics

### Architecture note

The compiler infrastructure (type system, lexer, parser, resolver, type
checker, LLVM type lowering) already supported i64 internally —
BuiltinKind::I64 was defined, lowered to LLVM i64, and handled in
is_numeric/is_integer. The gaps were only: runtime hooks, stdlib
extends, literal typing, and str_length narrowing.

## Test plan

- [x] All 10 test suites pass (including typecheck tests unchanged)
- [x] All 13 examples compile end-to-end
- [x] `i64.dao` produces correct output: factorial(20), generic min,
length
- [x] Golden AST files regenerated


---------

+++ runtime/core/string.c
@@ -35,9 +35,9 @@ struct dao_string __dao_str_concat(const struct dao_string *a,
   return (struct dao_string){.ptr = buf, .len = total};
 }
 
-int32_t __dao_str_length(const struct dao_string *s) {
+int64_t __dao_str_length(const struct dao_string *s) {
   if (s == NULL) {
     return 0;
   }
-  return (int32_t)s->len;
+  return s->len;
 }
